+++ src-ui/sdr/limesdr.cpp
@@ -115,7 +115,7 @@ void SDRLimeSDR::drawUI()
     ImGui::End();
 }
 
-void SDRLimeSDR::setFrequency(long frequency)
+void SDRLimeSDR::setFrequency(float frequency)
 {
     d_frequency = frequency;
     std::memcpy(this->frequency, std::to_string((float)d_frequency / 1e6).c_str(), std::to_string((float)d_frequency / 1e6).length());
