+++ src-ui/sdr/spyserver.cpp
@@ -116,7 +116,7 @@ void SDRSpyServer::drawUI()
     ImGui::End();
 }
 
-void SDRSpyServer::setFrequency(long frequency)
+void SDRSpyServer::setFrequency(float frequency)
 {
     d_frequency = frequency;
     std::memcpy(this->frequency, std::to_string((float)d_frequency / 1e6).c_str(), std::to_string((float)d_frequency / 1e6).length());
